requirements.txt
boto3>=1.34.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openai>=1.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
